+++ modules/output.nix
@@ -29,12 +29,6 @@ let
 in {
   options = {
 
-    systems = mkOption {
-      type = listOf str;
-      description = mdDoc "The systems for which to create outputs.";
-      default = ["x86_64-linux"];
-    };
-
     output = {
 
       extraPackages = mkOption {