+++ modules/system.nix
@@ -6,15 +6,10 @@ in {
   options = with types; {
 
     system = mkOption {
+      description = mdDoc "The system string like `x86_64-linux`, set by iterating over [](#opt-general-systems).";
       type = str;
-    };
-
-    systemOutputs = mkOption {
-      type = unspecified;
       readOnly = true;
     };
 
   };
-
-  config.systemOutputs = config.output.final;
 }